tests minus purging
